+++ src/models/gradle.ts
@@ -1,6 +1,6 @@
 export interface GradleDependency {
   name: string;
-  url: string;
+  url?: string;
   version: string;
   license: string;
   licenseUrl: string;
