+++ src/models/index.ts
@@ -1,11 +1,12 @@
 import { Configuration } from "./configuration.js";
-import { GradleDependency } from "./gradle.js";
+import { GradleDependency, GradleDependencyReport } from "./gradle.js";
 import { License } from "./license.js";
-import { MavenDependency, MavenLicense } from "./mavenDependency.js";
+import { MavenDependency, MavenLicense } from "./maven.js";
 
 export type {
   Configuration,
   GradleDependency,
+  GradleDependencyReport,
   License,
   MavenDependency,
   MavenLicense,
